webstorage/localstorage-about-blank-1P-iframe-partitioned.tentative.html

<!doctype html>
<meta charset=utf-8>
<title>localStorage: about:blank partitioning</title>
<meta name=help href="https://privacycg.github.io/storage-partitioning/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/webstorage/resources/partitioning-utils.js"></script>
<style>iframe {display: none}</style>
<body>
<script>
promise_test(t => {
  localStorage.clear();

  const id = getOrCreateID("userID2");

  t.add_cleanup(() => localStorage.clear());

  return addIframePromise("about:blank").then((iframe) => {
    const frameID = iframe.contentWindow.localStorage["userID2"];
    assert_false(typeof id === "undefined", "id is not undefined");
    assert_false(typeof frameID === "undefined", "id is not undefined");
    assert_true(id === frameID, "1P and about:blank iframe share StorageKey");
  });
}, "1P about:blank iframe shares StorageKey with 1P.");
</script>
</body>
